Refactoring: move ENVV methods in a ENVV::Base module
Clearer, improves modularity
